All excamples except fantasy sports use unified client style-ish. Other misc cleanup.
