--- examples/positionkeeper/client/positionkeeper/BenchmarkCallback.java
@@ -1,85 +1,0 @@
-/* This file is part of VoltDB.
- * Copyright (C) 2008-2016 VoltDB Inc.
- *
- * Permission is hereby granted, free of charge, to any person obtaining
- * a copy of this software and associated documentation files (the
- * "Software"), to deal in the Software without restriction, including
- * without limitation the rights to use, copy, modify, merge, publish,
- * distribute, sublicense, and/or sell copies of the Software, and to
- * permit persons to whom the Software is furnished to do so, subject to
- * the following conditions:
- *
- * The above copyright notice and this permission notice shall be
- * included in all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
- * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
- * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
- * IN NO EVENT SHALL THE AUTHORS BE LIABLE FOR ANY CLAIM, DAMAGES OR
- * OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE,
- * ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
- * OTHER DEALINGS IN THE SOFTWARE.
- */
-
-package positionkeeper;
-
-import org.voltdb.client.ClientResponse;
-import org.voltdb.client.ProcedureCallback;
-
-import com.google_voltpatches.common.collect.ConcurrentHashMultiset;
-import com.google_voltpatches.common.collect.Multiset;
-
-public class BenchmarkCallback implements ProcedureCallback {
-
-    private static Multiset<String> calls = ConcurrentHashMultiset.create();
-    private static Multiset<String> commits = ConcurrentHashMultiset.create();
-    private static Multiset<String> rollbacks = ConcurrentHashMultiset.create();
-
-    String procedureName;
-    long maxErrors;
-
-    public static void printProcedureResults(String procedureName) {
-        System.out.println("  " + procedureName);
-        System.out.println("        calls: " + calls.count(procedureName));
-        System.out.println("      commits: " + commits.count(procedureName));
-        System.out.println("    rollbacks: " + rollbacks.count(procedureName));
-    System.out.println();
-    }
-
-    public static void printAllResults() {
-    for (String e : calls.elementSet()) {
-        printProcedureResults(e);
-    }
-    }
-
-    public BenchmarkCallback(String procedure, long maxErrors) {
-        super();
-        this.procedureName = procedure;
-        this.maxErrors = maxErrors;
-    }
-
-    public BenchmarkCallback(String procedure) {
-        this(procedure, 5l);
-    }
-
-    @Override
-    public void clientCallback(ClientResponse cr) {
-
-    calls.add(procedureName,1);
-
-        if (cr.getStatus() == ClientResponse.SUCCESS) {
-        commits.add(procedureName,1);
-        } else {
-            long totalErrors = rollbacks.add(procedureName,1);
-
-            System.err.println("DATABASE ERROR: " + cr.getStatusString());
-
-            if (totalErrors > maxErrors) {
-                System.err.println("exceeded " + maxErrors + " maximum database errors - exiting client");
-                System.exit(-1);
-            }
-
-        }
-    }
-}
-
